third_party/WebKit/LayoutTests/fast/dom/MutationObserver/mutation-observer-prefix-expected.txt

Test that new WebKitMutationObserver is the same as new MutationObserver.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS noPrefixObserver instanceof WebKitMutationObserver is true
PASS prefixObserver instanceof WebKitMutationObserver is true
PASS noPrefixObserver instanceof MutationObserver is true
PASS prefixObserver instanceof MutationObserver is true
PASS successfullyParsed is true

TEST COMPLETE
